requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
arrow==1.2.3
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
base58==0.2.2
cffi==1.15.1
charset-normalizer==2.1.1
click==6.6
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
docker-py==1.8.0
ecdsa==0.18.0
exceptiongroup==1.0.4
flake8==6.0.0
frozenlist==1.3.3
funcsigs==1.0.2
future==0.18.2
idna==3.4
iniconfig==1.1.1
jsonrpcclient==2.0.1
jsonrpcserver==3.1.1
jsonschema==4.17.1
mccabe==0.7.0
mnemonic==0.13
multidict==6.0.2
packaging==21.3
path==16.5.0
path.py==12.5.0
pbkdf2==1.3
pexpect==4.8.0
pluggy==1.0.0
protobuf==3.0.0a3
ptyprocess==0.7.0
pyaes==1.6.1
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.16.0
pyflakes==3.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
python-binance-chain==0.1.5
python-dateutil==2.8.2
pytz==2022.6
pywallet==0.1.0
PyYAML==6.0
requests==2.11.1
secp256k1==0.14.0
sha256==0.3
six==1.16.0
sqlparse==0.4.3
tabulate==0.9.0
tomli==2.0.1
two1==3.10.9
tzdata==2022.6
websocket-client==1.4.2
websockets==10.4
yarl==1.8.1
